requirements.txt
discord.py
python-dotenv
pydantic
d20
aiohttp
python-dateutil